--- .idea/compiler.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="CompilerConfiguration">
-    <annotationProcessing>
-      <profile name="Gradle Imported" enabled="true">
-        <outputRelativeToContentRoot value="true" />
-        <processorPath useClasspath="false">
-          <entry name="$USER_HOME$/.gradle/caches/modules-2/files-2.1/org.projectlombok/lombok/1.18.38/57f8f5e02e92a30fd21b80cbd426a4172b5f8e29/lombok-1.18.38.jar" />
-        </processorPath>
-        <module name="likelion.sikjajaeDeal.backendProject.main" />
-      </profile>
-    </annotationProcessing>
-  </component>
-  <component name="JavacSettings">
-    <option name="ADDITIONAL_OPTIONS_OVERRIDE">
-      <module name="likelion.sikjajaeDeal.backendProject" options="-parameters" />
-      <module name="likelion.sikjajaeDeal.backendProject.main" options="-parameters" />
-      <module name="likelion.sikjajaeDeal.backendProject.test" options="-parameters" />
-    </option>
-  </component>
-</project>
