--- .idea/gradle.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="GradleSettings">
-    <option name="linkedExternalProjectsSettings">
-      <GradleProjectSettings>
-        <option name="externalProjectPath" value="$PROJECT_DIR$/beanBa" />
-        <option name="gradleJvm" value="homebrew-17" />
-        <option name="modules">
-          <set>
-            <option value="$PROJECT_DIR$/beanBa" />
-          </set>
-        </option>
-      </GradleProjectSettings>
-    </option>
-  </component>
-</project>
