+++ beanBa/src/main/java/likelion/beanBa/backendProject/BackendProjectApplication.java
@@ -2,6 +2,7 @@ package likelion.beanBa.backendProject;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 @SpringBootApplication
 public class BackendProjectApplication {
