+++ beanBa/src/main/java/likelion/beanBa/backendProject/member/security/config/SecurityConfig.java
@@ -49,9 +49,7 @@ public class SecurityConfig {
                         "/oauth2/**",
                         "/upload",
                         "/swagger-ui/**",
-                        "/v3/api-docs/**",
-                        "/api/test-sale-post/**",  //sale-post 테스트 하느라고 잠시 넣어놨습니다.
-                        "/api/sale-post/**"
+                        "/v3/api-docs/**"
                 ).permitAll()
                 .anyRequest().authenticated()
         );
