--- beanBa/src/main/java/likelion/beanBa/backendProject/product/entity/SaleCompleteRequest.java
@@ -1,8 +1,0 @@
-package likelion.beanBa.backendProject.product.entity;
-
-import lombok.Getter;
-
-@Getter
-public class SaleCompleteRequest {
-    private Long buyerPk; // 구매자 buy_member_pk
-}
