--- beanBa/src/main/java/likelion/beanBa/backendProject/product/product_enum/SaleStatement.java
@@ -1,17 +1,0 @@
-package likelion.beanBa.backendProject.product.product_enum;
-
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum SaleStatement {
-
-    S("S"),    // 판매중
-    C("C");// 판매완료
-
-    private final String code;
-
-
-}
